[AUTH] Setup auth services + migrate CAS into it + fix
FIX Error that force mentors to get their cas username as somethin.other
